fdsafea

+++ src/com/girlkun/jdbc/daos/GodGK.java
@@ -605,13 +605,13 @@ public class GodGK {
                                 pet.nPoint.tiemNang = Long.parseLong(String.valueOf(dataArray.get(2)));
                                 pet.nPoint.stamina = Short.parseShort(String.valueOf(dataArray.get(3)));
                                 pet.nPoint.maxStamina = Short.parseShort(String.valueOf(dataArray.get(4)));
-                                pet.nPoint.hpg = Integer.parseInt(String.valueOf(dataArray.get(5)));
-                                pet.nPoint.mpg = Integer.parseInt(String.valueOf(dataArray.get(6)));
-                                pet.nPoint.dameg = Integer.parseInt(String.valueOf(dataArray.get(7)));
-                                pet.nPoint.defg = Integer.parseInt(String.valueOf(dataArray.get(8)));
+                                pet.nPoint.hpg = Double.parseDouble(String.valueOf(dataArray.get(5)));
+                                pet.nPoint.mpg = Double.parseDouble(String.valueOf(dataArray.get(6)));
+                                pet.nPoint.dameg = Double.parseDouble(String.valueOf(dataArray.get(7)));
+                                pet.nPoint.defg = Double.parseDouble(String.valueOf(dataArray.get(8)));
                                 pet.nPoint.critg = Integer.parseInt(String.valueOf(dataArray.get(9)));
-                                int hp = Integer.parseInt(String.valueOf(dataArray.get(10)));
-                                int mp = Integer.parseInt(String.valueOf(dataArray.get(11)));
+                                double hp = Double.parseDouble(String.valueOf(dataArray.get(10)));
+                                double mp = Double.parseDouble(String.valueOf(dataArray.get(11)));
 
                                 //data body
                                 dataArray = (JSONArray) jv.parse(String.valueOf(petData.get(2)));
@@ -1120,8 +1120,8 @@ public class GodGK {
         try {
             rs = GirlkunDB.executeQuery("select * from player where id = ? limit 1", id);
             if (rs.first()) {
-                int plHp = 200000000;
-                int plMp = 200000000;
+                double plHp = 200000000;
+                double plMp = 200000000;
                 JSONValue jv = new JSONValue();
                 JSONArray dataArray = null;
 
@@ -1179,14 +1179,14 @@ public class GodGK {
                 player.nPoint.tiemNang = Long.parseLong(String.valueOf(dataArray.get(2)));
                 player.nPoint.stamina = Short.parseShort(String.valueOf(dataArray.get(3)));
                 player.nPoint.maxStamina = Short.parseShort(String.valueOf(dataArray.get(4)));
-                player.nPoint.hpg = Integer.parseInt(String.valueOf(dataArray.get(5)));
-                player.nPoint.mpg = Integer.parseInt(String.valueOf(dataArray.get(6)));
-                player.nPoint.dameg = Integer.parseInt(String.valueOf(dataArray.get(7)));
-                player.nPoint.defg = Integer.parseInt(String.valueOf(dataArray.get(8)));
+                player.nPoint.hpg = Double.parseDouble(String.valueOf(dataArray.get(5)));
+                player.nPoint.mpg = Double.parseDouble(String.valueOf(dataArray.get(6)));
+                player.nPoint.dameg = Double.parseDouble(String.valueOf(dataArray.get(7)));
+                player.nPoint.defg = Double.parseDouble(String.valueOf(dataArray.get(8)));
                 player.nPoint.critg = Byte.parseByte(String.valueOf(dataArray.get(9)));
                 dataArray.get(10); //** Năng động
-                plHp = Integer.parseInt(String.valueOf(dataArray.get(11)));
-                plMp = Integer.parseInt(String.valueOf(dataArray.get(12)));
+                plHp = Double.parseDouble(String.valueOf(dataArray.get(11)));
+                plMp = Double.parseDouble(String.valueOf(dataArray.get(12)));
                 dataArray.clear();
 
                 //data đậu thần
